+++ services/reventService.tsx
@@ -101,6 +101,7 @@ export async function getCollectionList() {
 export async function getContestList() {
   const contests = [
     {
+      id: 1,
       name: 'Coding Challenge #1',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -109,6 +110,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 2,
       name: 'Coding Challenge #2',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -117,6 +119,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 3,
       name: 'Coding Challenge #3',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -125,6 +128,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 4,
       name: 'Coding Challenge #4',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -133,6 +137,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 5,
       name: 'Coding Challenge #5',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -141,6 +146,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 6,
       name: 'Coding Challenge #6',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -149,6 +155,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 7,
       name: 'Coding Challenge #7',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -157,6 +164,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 8,
       name: 'Coding Challenge #8',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -165,6 +173,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 9,
       name: 'Coding Challenge #9',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -173,6 +182,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 10,
       name: 'Coding Challenge #10',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -181,6 +191,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 11,
       name: 'Design Contest #2',
       description:
         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
@@ -189,6 +200,7 @@ export async function getContestList() {
       status: 'Voting',
     },
     {
+      id: 12,
       name: 'Hackathon #1',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2023-08-01T09:00:00Z',
@@ -196,6 +208,7 @@ export async function getContestList() {
       status: 'Closed',
     },
     {
+      id: 13,
       name: 'Coding Challenge #2',
       description: 'A coding challenge to test your skills',
       dateStart: '2023-09-01T09:00:00Z',
@@ -203,6 +216,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 14,
       name: 'Design Contest #2',
       description: 'A design contest to showcase your creativity',
       dateStart: '2023-10-01T09:00:00Z',
@@ -210,6 +224,7 @@ export async function getContestList() {
       status: 'Closed',
     },
     {
+      id: 15,
       name: 'Hackathon #2',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2023-11-01T09:00:00Z',
@@ -217,6 +232,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 16,
       name: 'Coding Challenge #3',
       description: 'A coding challenge to test your skills',
       dateStart: '2024-01-01T09:00:00Z',
@@ -224,6 +240,7 @@ export async function getContestList() {
       status: 'Voting',
     },
     {
+      id: 17,
       name: 'Design Contest #3',
       description: 'A design contest to showcase your creativity',
       dateStart: '2024-02-01T09:00:00Z',
@@ -231,6 +248,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 18,
       name: 'Hackathon #3',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2024-03-01T09:00:00Z',
@@ -238,6 +256,7 @@ export async function getContestList() {
       status: 'Closed',
     },
     {
+      id: 19,
       name: 'Coding Challenge #4',
       description: 'A coding challenge to test your skills',
       dateStart: '2024-04-01T09:00:00Z',
@@ -245,6 +264,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 20,
       name: 'Coding Challenge #1',
       description: 'A coding challenge to test your skills',
       dateStart: '2023-06-01T09:00:00Z',
@@ -252,6 +272,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 21,
       name: 'Design Contest #1',
       description: 'A design contest to showcase your creativity',
       dateStart: '2023-07-01T09:00:00Z',
@@ -259,6 +280,7 @@ export async function getContestList() {
       status: 'Voting',
     },
     {
+      id: 22,
       name: 'Hackathon #1',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2023-08-01T09:00:00Z',
@@ -266,6 +288,7 @@ export async function getContestList() {
       status: 'Closed',
     },
     {
+      id: 23,
       name: 'Coding Challenge #2',
       description: 'A coding challenge to test your skills',
       dateStart: '2023-09-01T09:00:00Z',
@@ -273,6 +296,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 24,
       name: 'Design Contest #2',
       description: 'A design contest to showcase your creativity',
       dateStart: '2023-10-01T09:00:00Z',
@@ -280,6 +304,7 @@ export async function getContestList() {
       status: 'Closed',
     },
     {
+      id: 25,
       name: 'Hackathon #2',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2023-11-01T09:00:00Z',
@@ -287,6 +312,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 26,
       name: 'Coding Challenge #3',
       description: 'A coding challenge to test your skills',
       dateStart: '2024-01-01T09:00:00Z',
@@ -294,6 +320,7 @@ export async function getContestList() {
       status: 'Voting',
     },
     {
+      id: 27,
       name: 'Design Contest #3',
       description: 'A design contest to showcase your creativity',
       dateStart: '2024-02-01T09:00:00Z',
@@ -301,6 +328,7 @@ export async function getContestList() {
       status: 'Open',
     },
     {
+      id: 28,
       name: 'Hackathon #3',
       description: 'A hackathon to develop innovative solutions',
       dateStart: '2024-03-01T09:00:00Z',
